pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "replaypack"
version = "0.1.0"
description = "Replay Pack CLI for agent log capture, redaction, and MCP replay"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "ReplayPack Contributors" }]
dependencies = [
  "pydantic>=2.8,<3",
  "typer>=0.12,<1",
  "PyYAML>=6.0",
  "httpx>=0.27,<1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "mypy>=1.11",
  "ruff>=0.6",
  "pre-commit>=3.8",
]

[project.scripts]
replaypack = "replaypack.cli:main"

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "W"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
show_error_codes = true
pretty = true
exclude = ["^tests/fixtures/"]

[[tool.mypy.overrides]]
module = ["yaml"]
ignore_missing_imports = true
